+++ tests/source/configs/use_field_init_shorthand/false.rs
@@ -1,3 +1,4 @@
+// rustfmt-use_field_init_shorthand: false
 // Use field initialization shorthand if possible.
 
 fn main() {