+++ tests/source/configs/use_field_init_shorthand/true.rs
@@ -1,13 +1,18 @@
+// rustfmt-use_field_init_shorthand: true
 // Use field initialization shorthand if possible.
 
 fn main() {
-    let a = Foo { x, y, z };
+    let a = Foo {
+        x: x,
+        y: y,
+        z: z,
+    };
 
     let b = Bar {
-        x,
-        y,
+        x: x,
+        y: y,
         #[attr]
-        z,
+        z: z,
         #[rustfmt_skip]
         skipped: skipped,
     };